Исправлен баг при котором имя публициста не отображалось Изменены настройки связи между news и publicist

+++ src/main/java/com/example/ilsan34/clientManager/service/PublicistService.java
@@ -2,16 +2,17 @@ package com.example.ilsan34.clientManager.service;
 
 import com.example.ilsan34.clientManager.model.Publicist;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface PublicistService {
     void addNewPublicist(Publicist publicist);
 
-    Publicist getById(Long id);
+    Publicist getById(BigInteger id);
 
     void updatePublicist(Publicist publicist);
 
-    void deletePublicist(Long id);
+    void deletePublicist(BigInteger id);
 
     List<Publicist> getAll();
 }
